requirements.txt
termcolor==1.1.0
torchvision==0.9.1
tqdm==4.61.1
albumentations==0.5.1
numpy==1.19.2
torch==1.8.1
absl_py==0.13.0
opencv_contrib_python==4.5.1.48
Pillow==8.2.0
